Sync upstream evo-design/proto-tools main (#157)
* feat(cli): add `signature` for payload-free symbol discovery (#37)

agent_context.md tells agents not to guess symbol names from the registry key,
and until now the only surface that answered that question was
`example-input --as-python`. That command carries the example's real values, so
asking it for four symbol names costs whatever the example weighs: 368 KB for
ipsae-scoring (a barnase-barstar PDB plus its PAE matrix), 524 KB for borzoi.
Nothing else exposes the run-function name or the import module; `input`,
`config`, `output`, `schema`, and `list --json` all omit both.

`signature` answers just that question:

    from proto_tools.tools.sequence_alignment.blast.create_blast_db import (
        CreateBlastDbConfig,
        CreateBlastDbInput,
        run_create_blast_db,
    )

    result = run_create_blast_db(
        CreateBlastDbInput(fasta=...),
        CreateBlastDbConfig(),  # optional, omit for defaults
    )  # -> CreateBlastDbOutput

Nothing in the output scales with the example, so the cost is flat: across all
133 registered tools the median is 383 bytes and the maximum is 504, against
4.8 MB total for the example-input JSON path.

Notes on the shape:

- Output is named in the trailing annotation but not imported. Callers never
  construct one, so importing it would leave an unused name in pasted code, and
  for 24 tools it lives in a different module from the Input.
- `...` in argument position is deliberately not runnable. An elided literal
  that looks runnable and then fails validation costs a retry to discover;
  `...` cannot be mistaken for a value.
- Imports are grouped by defining module (at most two across all tools) rather
  than by toolkit package, so they do not depend on `__init__` re-exports.
- Only required input fields are listed; `input <tool>` remains the field-level
  detail view.

`--json` emits the same symbols for MCP servers and other programmatic callers.

example-input keeps its current behavior and stays the way to get a payload you
can actually execute; it is simply no longer the documented path for learning
names. Its payload sizes are largely inherent rather than fixture bloat, since
BorzoiInput requires exactly 524,288 bp and the structure fixtures back
integration tests.

* Make esmc-sae-features a tool of the esmc toolkit (#36)

* Make esmc-sae-features a tool of the esmc toolkit

* Fold the SAE-selection point into the preceding README paragraph

* Unbold the SAE-selection point in the notebook, matching the README

* fix(pyrosetta): call set_interface through DockingPartners on current builds (#38)

Fixes #24
