+++ proto_tools/tools/masked_models/__init__.py
@@ -54,14 +54,12 @@ from proto_tools.tools.masked_models.esmc import (
     ESMCEmbeddingsConfig,
     ESMCEmbeddingsInput,
     ESMCEmbeddingsOutput,
-    run_esmc_embeddings,
-)
-from proto_tools.tools.masked_models.esmc_sae import (
     ESMCSAEFeaturesConfig,
     ESMCSAEFeaturesInput,
     ESMCSAEFeaturesOutput,
     SAELayerFeatures,
     SequenceSAEFeatures,
+    run_esmc_embeddings,
     run_esmc_sae_features,
 )
 from proto_tools.tools.masked_models.shared_data_models import (
