+++ proto_tools/tools/masked_models/esmc/examples/example_output/esmc_sae_features.csv
