--- proto_tools/tools/masked_models/esmc_sae/cite.bib
@@ -1,16 +1,0 @@
-@misc{candido2026language,
-  title={Language Modeling Materializes a World Model of Protein Biology},
-  author={Candido, Salvatore and Hayes, Thomas and Derry, Alexander and Rao, Roshan and Lin, Zeming and Verkuil, Robert and Wu, Bryan and Lee, Jin Sub and Bruguera, Elise S. and Keval, Jehan A. and Kopylov, Mykhailo and Pak, John E. and Wu, Wesley and Thomas, Neil and Mataraso, Samson and Hsu, Alvin and Trotman-Grant, Ashton C. and Fatras, Kilian and dos Santos Costa, Allan and Badkundri, Rohil and Ak{\i}n, Halil and Oktay, Deniz and Deaton, Jonathan and Montabana, Elizabeth and Sitwala, Hrishita and Yu, Yue and Wiggert, Marius and Carlin, Dylan Alexander and Goering, Anthony W. and Blazejewski, Tomasz and Sandora, McCullen and Hla, Michael and Jia, Tina Z. and Kloker, Leon H. and Sofroniew, Nicholas J. and Uehara, Masatoshi and Pannu, Jassi and Bachas, Sharrol and Liu, Daniel S. and Sercu, Tom and Rives, Alexander},
-  year={2026},
-  url={https://www.biorxiv.org/content/10.64898/2026.06.03.729735},
-  note={Preprint}
-}
-
-@software{evolutionaryscale_2024,
-  author={{EvolutionaryScale Team}},
-  title={evolutionaryscale/esm},
-  year={2024},
-  publisher={Zenodo},
-  doi={10.5281/zenodo.14219303},
-  url={https://doi.org/10.5281/zenodo.14219303}
-}
