--- proto_tools/tools/masked_models/esmc_sae/license.yaml
@@ -1,11 +1,0 @@
-code:
-  spdx: MIT
-  url: https://github.com/Biohub/esm/blob/main/LICENSE.md
-weights:
-  spdx: MIT
-  url: https://huggingface.co/biohub/ESMC-SAE-Overview
-commercial_use: 'yes'
-redistribution: true
-attribution_required: false
-notes: ESM C SAE code and HuggingFace weights are MIT-licensed under Chan Zuckerberg Biohub copyright, matching the ESM C backbones they decompose. The model cards additionally tag 'other', which links only to a third-party dependency notice imposing no terms on users beyond those dependencies' own licenses.
-last_updated: '2026-07-29'
