--- proto_tools/tools/masked_models/esmc_sae/links.yaml
@@ -1,6 +1,0 @@
-github: https://github.com/Biohub/esm
-website: https://biohub.ai/
-paper: https://www.biorxiv.org/content/10.64898/2026.06.03.729735
-huggingface: https://huggingface.co/biohub/ESMC-SAE-Overview
-organizations:
-  - Biohub
