--- proto_tools/tools/masked_models/esmc_sae/standalone/shared_env.txt
@@ -1,1 +1,0 @@
-biohub_esm
