+++ tests/tool_infra_tests/test_cli.py
@@ -134,6 +134,100 @@ def test_unknown_section_exits_one() -> None:
     assert "not found" in err
 
 
+# ── Signature ──────────────────────────────────────────────────────────────
+
+
+def test_signature_names_symbols_the_key_cannot_predict() -> None:
+    """The point of the verb: registry key 'blast-create-db' does not yield these names."""
+    code, out, _ = _run("signature", "blast-create-db")
+    assert code == 0
+    assert "CreateBlastDbInput" in out  # not BlastCreateDbInput
+    assert "run_create_blast_db" in out  # not run_blast_create_db
+    assert "CreateBlastDbConfig" in out
+    assert "CreateBlastDbOutput" in out
+    assert "proto_tools.tools.sequence_alignment.blast.create_blast_db" in out
+
+
+def test_signature_names_the_output_without_importing_it() -> None:
+    """Callers never construct an Output, so importing it would leave an unused name."""
+    code, out, _ = _run("signature", "ipsae-scoring")
+    assert code == 0
+    imports, call = out.split("\n\nresult =", maxsplit=1)
+    assert "IPSAEScoringOutput" not in imports
+    assert "-> IPSAEScoringOutput" in call
+
+
+def test_signature_lists_required_input_fields_only() -> None:
+    code, out, _ = _run("signature", "ipsae-scoring")
+    assert code == 0
+    assert "IPSAEScoringInput(structure=..., binder_chain=..., target_chains=...)" in out
+
+
+def test_signature_resolves_a_run_function_name() -> None:
+    code, out, _ = _run("signature", "run_create_blast_db")
+    assert code == 0
+    assert "CreateBlastDbInput" in out
+
+
+def test_signature_stays_small_for_a_tool_with_a_huge_example() -> None:
+    """The reason the verb exists: cost is fixed, not proportional to the example payload.
+
+    borzoi's example input is a 524,288 bp window (the model context length, enforced by
+    ``BorzoiInput``), so ``example-input`` cannot be made cheap for it at any fixture size.
+    """
+    code, signature, _ = _run("signature", "borzoi-prediction")
+    assert code == 0
+    _, example, _ = _run("example-input", "borzoi-prediction")
+    assert len(signature) < 1024
+    assert len(example) > 500_000
+
+
+def test_signature_json_carries_symbols_and_modules() -> None:
+    code, out, _ = _run("signature", "blast-create-db", "--json")
+    assert code == 0
+    payload = json.loads(out)
+    assert payload["input_class"] == "CreateBlastDbInput"
+    assert payload["run_function"] == "run_create_blast_db"
+    assert payload["output_class"] == "CreateBlastDbOutput"
+    assert payload["required_input_fields"] == ["fasta"]
+    assert payload["modules"]["input_class"].startswith("proto_tools.tools.")
+
+
+@pytest.mark.extensive
+def test_signature_parses_and_imports_resolve_for_every_tool() -> None:
+    """Every rendered signature is valid Python whose imports bind the names it uses."""
+    import ast
+
+    from proto_tools.tools.tool_registry import ToolRegistry
+
+    specs = {s.key: s for s in [*ToolRegistry.list_cpu_tools(), *ToolRegistry.list_gpu_tools()]}
+    failures: list[str] = []
+    for key, spec in sorted(specs.items()):
+        code, out, _ = _run("signature", key)
+        if code != 0:
+            failures.append(f"{key}: exit {code}")
+            continue
+        try:
+            tree = ast.parse(out)
+        except SyntaxError as exc:
+            failures.append(f"{key}: {exc}")
+            continue
+        imports = [node for node in tree.body if isinstance(node, ast.Import | ast.ImportFrom)]
+        namespace: dict[str, object] = {}
+        try:
+            exec(compile(ast.Module(body=imports, type_ignores=[]), "<generated>", "exec"), namespace)  # noqa: S102
+        except Exception as exc:
+            failures.append(f"{key}: {type(exc).__name__}: {exc}")
+            continue
+        failures.extend(
+            f"{key}: emitted import did not bind {symbol}"
+            for symbol in (spec.input_model.__name__, spec.config_model.__name__, spec.function.__name__)
+            if symbol not in namespace
+        )
+
+    assert not failures, "signatures with unusable imports:\n" + "\n".join(failures)
+
+
 # ── Schema / example-input ─────────────────────────────────────────────────
